# api/configure/server/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Server API
    version: 0.0.1
paths:
    /configure/api/v1/server:
        put:
            tags:
                - Server
            operationId: Server_UpdateServer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/configure.api.configure.server.v1.UpdateServerRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/configure.api.configure.server.v1.UpdateServerReply'
        post:
            tags:
                - Server
            description: CreateServer 创建服务
            operationId: Server_CreateServer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/configure.api.configure.server.v1.CreateServerRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/configure.api.configure.server.v1.CreateServerReply'
        delete:
            tags:
                - Server
            operationId: Server_DeleteServer
            parameters:
                - name: id
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/configure.api.configure.server.v1.DeleteServerReply'
    /configure/api/v1/servers:
        get:
            tags:
                - Server
            description: ListServer 获取服务信息列表
            operationId: Server_ListServer
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: order
                  in: query
                  schema:
                    type: string
                - name: orderBy
                  in: query
                  schema:
                    type: string
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: name
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: boolean
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/configure.api.configure.server.v1.ListServerReply'
components:
    schemas:
        configure.api.configure.server.v1.CreateServerReply:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
        configure.api.configure.server.v1.CreateServerRequest:
            type: object
            properties:
                keyword:
                    type: string
                name:
                    type: string
                description:
                    type: string
                status:
                    type: boolean
        configure.api.configure.server.v1.DeleteServerReply:
            type: object
            properties: {}
        configure.api.configure.server.v1.ListServerReply:
            type: object
            properties:
                total:
                    type: integer
                    format: uint32
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/configure.api.configure.server.v1.ListServerReply_Server'
        configure.api.configure.server.v1.ListServerReply_Server:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                keyword:
                    type: string
                name:
                    type: string
                description:
                    type: string
                status:
                    type: boolean
                createdAt:
                    type: integer
                    format: uint32
                updatedAt:
                    type: integer
                    format: uint32
        configure.api.configure.server.v1.UpdateServerReply:
            type: object
            properties: {}
        configure.api.configure.server.v1.UpdateServerRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                keyword:
                    type: string
                name:
                    type: string
                status:
                    type: boolean
                description:
                    type: string
tags:
    - name: Server